requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.52.0
django-cors-headers==3.13.0
django-cors-middleware==1.5.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
idna==3.4
oauthlib==3.2.2
psycopg2==2.9.5
pycparser==2.21
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.3.0
